requirements.txt
numpy>=1.22.3
matplotlib>=3.5.1
pydicom>=2.3.0
requests>=2.28.1
pillow>=9.2.0
scikit-image>=0.19.2
ango>=0.0.21